feat(activities): made the recent activities reusable
