+++ app/dashboard/page.tsx
@@ -38,6 +38,26 @@ const stats = [
         icon: ChartColumn,
     }
 ];
+const activities = [
+    {
+        id: "!",
+        type: 'feedback',
+        title: 'Lisa Thompson provided feedback for Michael Rodriguez',
+        timestamp: '13/02/2024, 12:15:00'
+    },
+    {
+        id: "2",
+        type: 'self-assessment',
+        title: 'Michael Rodriguez submitted a self-assessment',
+        timestamp: '13/02/2024, 12:15:00'
+    },
+    {
+        id: "3",
+        type: 'ai-score',
+        title: 'AI score updated for Jane Doe',
+        timestamp: '13/02/2024, 12:15:00'
+    }
+];
 
 export default function Dashboard() {
     return (
@@ -51,7 +71,7 @@ export default function Dashboard() {
                 }
             </section>
             <section className="grid grid-cols-1 lg:grid-cols-2 gap-6">
-                <RecentActivity />
+                <RecentActivity activities={activities}/>
                 <AiInsights />
             </section>
         </main>
