+++ lib/types/activities.ts
@@ -1,0 +1,6 @@
+export interface ActivityPayload {
+    id: string;
+    type: string;
+    title: string;
+    timestamp: string;
+}
